requirements.txt
# TDnet Financial Intelligence Platform Requirements

SQLAlchemy>=2.0.0
alembic>=1.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
greenlet==3.2.2
idna==3.10
numpy==2.2.5
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pypdfium2==4.30.1
python-dateutil>=2.8.2
python-dotenv==0.21.0
pytz==2025.2
requests>=2.31.0
schedule==1.2.0
selenium>=4.15.0
six==1.17.0
soupsieve==2.7
tabulate>=0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager>=4.0.1
sentence-transformers>=2.2.0
transformers>=4.30.0
faiss-cpu>=1.7.0
scikit-learn>=1.3.0
asyncpg>=0.28.0
tqdm>=4.65.0
huggingface-hub>=0.15.0
migra>=1.1.0
lxml>=6.0.0
xlrd>=2.0.1